+++ com.github.eclipsecolortheme.test/src/com/github/eclipsecolortheme/test/mock/MockEclipsePreferences.java
@@ -12,12 +12,10 @@ public class MockEclipsePreferences implements IEclipsePreferences {
 
 	private Map<String, String> store = new HashMap<String, String>();
 	
-	@Override
 	public void put(String key, String value) {
 		store.put(key, value);
 	}
 
-	@Override
 	public String get(String key, String def) {
 		if (store.containsKey(key)) {
 			return store.get(key);
@@ -26,173 +24,144 @@ public class MockEclipsePreferences implements IEclipsePreferences {
 		}
 	}
 
-	@Override
 	public void remove(String key) {
 		store.remove(key);
 	}
 
-	@Override
 	public void clear() throws BackingStoreException {
 		store.clear();
 	}
 
-	@Override
 	public void putInt(String key, int value) {
 		store.put(key, String.valueOf(value));
 	}
 
-	@Override
 	public int getInt(String key, int def) {
 		// TODO Auto-generated method stub
 		return 0;
 	}
 
-	@Override
 	public void putLong(String key, long value) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public long getLong(String key, long def) {
 		// TODO Auto-generated method stub
 		return 0;
 	}
 
-	@Override
 	public void putBoolean(String key, boolean value) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public boolean getBoolean(String key, boolean def) {
 		// TODO Auto-generated method stub
 		return false;
 	}
 
-	@Override
 	public void putFloat(String key, float value) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public float getFloat(String key, float def) {
 		// TODO Auto-generated method stub
 		return 0;
 	}
 
-	@Override
 	public void putDouble(String key, double value) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public double getDouble(String key, double def) {
 		// TODO Auto-generated method stub
 		return 0;
 	}
 
-	@Override
 	public void putByteArray(String key, byte[] value) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public byte[] getByteArray(String key, byte[] def) {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public String[] keys() throws BackingStoreException {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public String[] childrenNames() throws BackingStoreException {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public Preferences parent() {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public boolean nodeExists(String pathName) throws BackingStoreException {
 		// TODO Auto-generated method stub
 		return false;
 	}
 
-	@Override
 	public String name() {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public String absolutePath() {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public void flush() throws BackingStoreException {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void sync() throws BackingStoreException {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void addNodeChangeListener(INodeChangeListener listener) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void removeNodeChangeListener(INodeChangeListener listener) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void addPreferenceChangeListener(IPreferenceChangeListener listener) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void removePreferenceChangeListener(
 			IPreferenceChangeListener listener) {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public void removeNode() throws BackingStoreException {
 		// TODO Auto-generated method stub
 
 	}
 
-	@Override
 	public Preferences node(String path) {
 		// TODO Auto-generated method stub
 		return null;
 	}
 
-	@Override
 	public void accept(IPreferenceNodeVisitor visitor)
 			throws BackingStoreException {
 		// TODO Auto-generated method stub